refactor: main.js 코드의 구조가 복잡해진 것을 파일을 나누어 가독성있게 구조 수정, 매직 넘버 수정
