+++ app.js
@@ -1,14 +1,17 @@
 const express = require('express')
+const path = require('path')
+
 const app = express()
 const PORT = 3000
 
-app.listen(PORT, function () {})
+app.get('/', (req, res) => {
+  res.sendFile(path.join(__dirname, '/index.html'))
+})
 
-// app.use(express.static('./'))
-app.use(express.static('css'))
-app.use(express.static('js'))
-app.use(express.static('images'))
+app.use('/', express.static(path.join(__dirname, './js')))
+app.use('/', express.static(path.join(__dirname, './css')))
+app.use('/', express.static(path.join(__dirname, './images')))
 
-app.get('/', function (req, res) {
-  res.sendFile(__dirname + '/index.html')
+app.listen(PORT, () => {
+  console.log(`App running on port ${PORT}`)
 })
