+++ index.html
@@ -22,7 +22,7 @@
     <script src="main.js" type="module"></script>
   </head>
   <body>
-    <div class="container">
+    <div id="container">
       <!-- header -->
       <header class="header">
         <div class="container">
@@ -129,58 +129,110 @@
                   alt="웹툰"
                   width="720px"
                 />
-              </div>
-              <div class="webtoon-slide__btn">
-                <div
-                  class="webtoon-slide__btn--back webtoon-slide__btn--common"
-                >
-                  <img src="./ic-banner-paging-back-nor.svg" alt="이전" />
+                <div class="webtoon-slide__desc">
+                  <h3 class="slide-title">화타가 된 외과의사</h3>
+                  <ul class="slide-desc__list">
+                    <li class="slide-desc__item">
+                      <div class="slide-desc__item--event">
+                        <img
+                          src="./badge-bigthum-event.svg"
+                          alt="이벤트"
+                          height="20px"
+                        />
+                      </div>
+                    </li>
+                    <li class="slide-desc__item">
+                      <div class="slide-desc__clock">
+                        <img
+                          src="./ico-bigthum-wait.svg"
+                          alt="기다리면 무료"
+                          height="20px"
+                        />
+                        <div class="slide-desc__item--desc">웹툰</div>
+                      </div>
+                    </li>
+                    <li class="slide-desc__item">
+                      <img
+                        src="./line_top_banner.png"
+                        alt="bar"
+                        height="20px"
+                      />
+                    </li>
+                    <li class="slide-desc__item">
+                      <div class="slide-desc__cnt">
+                        <img
+                          src="./ico-bigthum-person.svg"
+                          alt="ico"
+                          height="20px"
+                        />
+                        <div class="slide-desc__item--desc">27.6만명</div>
+                      </div>
+                    </li>
+                  </ul>
                 </div>
-                <div
-                  class="webtoon-slide__btn--next webtoon-slide__btn--common"
-                >
-                  <img src="./ic-banner-paging-next-nor.svg" alt="다음" />
+                <div class="webtoon-slide__banner">
+                  웹툰&소설 정주행, 캐시 이벤트!
                 </div>
               </div>
-              <div class="webtoon-slide__desc">
-                <h3 class="slide-title">화타가 된 외과의사</h3>
-                <ul class="slide-desc__list">
-                  <li class="slide-desc__item">
-                    <div class="slide-desc__item--event">
+              <div class="webtoon-slide__imgs">
+                <img
+                  class="webtoon-slide__imgs--first"
+                  src="./resource.png"
+                  alt="웹툰"
+                  width="720px"
+                />
+                <div class="webtoon-slide__desc">
+                  <h3 class="slide-title">화타가 된 외과의사</h3>
+                  <ul class="slide-desc__list">
+                    <li class="slide-desc__item">
+                      <div class="slide-desc__item--event">
+                        <img
+                          src="./badge-bigthum-event.svg"
+                          alt="이벤트"
+                          height="20px"
+                        />
+                      </div>
+                    </li>
+                    <li class="slide-desc__item">
+                      <div class="slide-desc__clock">
+                        <img
+                          src="./ico-bigthum-wait.svg"
+                          alt="기다리면 무료"
+                          height="20px"
+                        />
+                        <div class="slide-desc__item--desc">웹툰</div>
+                      </div>
+                    </li>
+                    <li class="slide-desc__item">
                       <img
-                        src="./badge-bigthum-event.svg"
-                        alt="이벤트"
+                        src="./line_top_banner.png"
+                        alt="bar"
                         height="20px"
                       />
-                    </div>
-                  </li>
-                  <li class="slide-desc__item">
-                    <div class="slide-desc__clock">
-                      <img
-                        src="./ico-bigthum-wait.svg"
-                        alt="기다리면 무료"
-                        height="20px"
-                      />
-                      <div class="slide-desc__item--desc">웹툰</div>
-                    </div>
-                  </li>
-                  <li class="slide-desc__item">
-                    <img src="./line_top_banner.png" alt="bar" height="20px" />
-                  </li>
-                  <li class="slide-desc__item">
-                    <div class="slide-desc__cnt">
-                      <img
-                        src="./ico-bigthum-person.svg"
-                        alt="ico"
-                        height="20px"
-                      />
-                      <div class="slide-desc__item--desc">27.6만명</div>
-                    </div>
-                  </li>
-                </ul>
+                    </li>
+                    <li class="slide-desc__item">
+                      <div class="slide-desc__cnt">
+                        <img
+                          src="./ico-bigthum-person.svg"
+                          alt="ico"
+                          height="20px"
+                        />
+                        <div class="slide-desc__item--desc">27.6만명</div>
+                      </div>
+                    </li>
+                  </ul>
+                </div>
+                <div class="webtoon-slide__banner">
+                  웹툰&소설 정주행, 캐시 이벤트!
+                </div>
               </div>
-              <div class="webtoon-slide__banner">
-                웹툰&소설 정주행, 캐시 이벤트!
+            </div>
+            <div class="webtoon-slide__btn">
+              <div class="webtoon-slide__btn--back webtoon-slide__btn--common">
+                <img src="./ic-banner-paging-back-nor.svg" alt="이전" />
+              </div>
+              <div class="webtoon-slide__btn--next webtoon-slide__btn--common">
+                <img src="./ic-banner-paging-next-nor.svg" alt="다음" />
               </div>
             </div>
           </section>
