+++ js/slide.js
@@ -1,0 +1,7 @@
+function setSlideImg(root, el) {
+  const elCopy = el.cloneNode(true)
+
+  root.appendChild(elCopy)
+}
+
+export default { setSlideImg }
