+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "node-sass": "node-sass",
-    "sass": "node-sass -wr scss/style.scss ./style.css"
+    "sass": "sass --watch ./scss/style.scss ./css/style.css",
+    "nodemon": "nodemon ./app.js",
+    "dev": "concurrently \"npm run sass\" \"npm run nodemon\""
   },
   "repository": {
     "type": "git",
@@ -20,7 +21,8 @@
   },
   "homepage": "https://github.com/uBeom/fe-kakaopage#readme",
   "dependencies": {
+    "concurrently": "^7.0.0",
     "express": "^4.17.3",
-    "node-sass": "^7.0.1"
+    "sass": "^1.49.8"
   }
 }
